project-tonixsmm/ProjectPart4/requirements.txt
mysql-connector-python==8.0.32
mysql-connector-python==8.0.32
Flask==2.2.2